requirements.txt
google-cloud-pubsub==2.11.0
avro==1.11.0
requests
python-dateutil
google-cloud-bigquery
google-cloud-bigtable
google-cloud-storage
google-auth

